fix(api): move fhir consolidated to sdk
Ref. metriport/metriport-internal#1158
