+++ packages/api/src/command/medical/patient/consolidated-get.ts
@@ -8,10 +8,12 @@ import {
   ResourceType,
 } from "@medplum/fhirtypes";
 import { ResourceTypeForConsolidation } from "@metriport/api-sdk";
-import { ConsolidationConversionType } from "@metriport/core/src/domain/fhir-to-medical-record";
 import { MetriportError } from "@metriport/core/util/error/metriport-error";
 import { makeLambdaClient } from "@metriport/core/external/aws/lambda";
-import { FhirToMedicalRecordPayload } from "@metriport/core/src/domain/fhir-to-medical-record";
+import {
+  FhirToMedicalRecordPayload,
+  ConsolidationConversionType,
+} from "@metriport/api-sdk/medical/models/fhir";
 import { Config } from "../../../shared/config";
 import { Patient } from "../../../domain/medical/patient";
 import { QueryProgress } from "../../../domain/medical/query-status";
